Update grammar for structure loading performance (see Readme)
This new grammar solves the fact that, most of the time, when loading an
octree structure, we want to know the bounding boxes of the nodes right
away, and the size of the data they represent. This is information we
want to know about unloaded nodes, so it shouldn't be at the same place at the
data. Before, being stored at the same place as the data, this
information needed back-and-forth reads in the octree file (from
STRUCTURE to DATA), which was a major performance concern.